--- src/phaser/enums.ts
@@ -1,12 +1,0 @@
-export const Events = {
-  SET_COMMANDS: "set-commands",
-  SET_LEVEL_ID: "set-level-id",
-  GAMEPLAY_SCENE_READY: "gameplay-scene-ready",
-} as const
-export type Event = (typeof Events)[keyof typeof Events]
-
-export const Variables = {
-  COMMANDS: "commands",
-  LEVEL_ID: "levelId",
-} as const
-export type Variable = (typeof Variables)[keyof typeof Variables]
